+++ lib/src/client.dart
@@ -43,7 +43,7 @@ import 'package:matrix/src/utils/try_get_push_rule.dart';
 
 typedef RoomSorter = int Function(Room a, Room b);
 
-enum LoginState { loggedIn, loggedOut }
+enum LoginState { loggedIn, loggedOut, softLoggedOut }
 
 extension TrailingSlash on Uri {
   Uri stripTrailingSlash() => path.endsWith('/')
@@ -88,6 +88,10 @@ class Client extends MatrixApi {
   final bool mxidLocalPartFallback;
 
   bool shareKeysWithUnverifiedDevices;
+
+  Future<void> Function(Client client)? onSoftLogout;
+
+  DateTime? accessTokenExpiresAt;
 
   // For CommandsClientExtension
   final Map<String, FutureOr<String?> Function(CommandArgs)> commands = {};
@@ -161,6 +165,12 @@ class Client extends MatrixApi {
     Set<KeyVerificationMethod>? verificationMethods,
     http.Client? httpClient,
     Set<String>? importantStateEvents,
+
+    /// You probably don't want to add state events which are also
+    /// in important state events to this list, or get ready to face
+    /// only having one event of that particular type in preLoad because
+    /// previewEvents are stored with stateKey '' not the actual state key
+    /// of your state event
     Set<String>? roomPreviewLastEvents,
     this.pinUnreadRooms = false,
     this.pinInvitedRooms = true,
@@ -178,6 +188,13 @@ class Client extends MatrixApi {
     this.shareKeysWithUnverifiedDevices = true,
     this.enableDehydratedDevices = false,
     this.receiptsPublicByDefault = true,
+
+    /// Implement your https://spec.matrix.org/v1.9/client-server-api/#soft-logout
+    /// logic here.
+    /// Set this to `refreshAccessToken()` for the easiest way to handle the
+    /// most common reason for soft logouts.
+    /// You can also perform a new login here by passing the existing deviceId.
+    this.onSoftLogout,
   })  : syncFilter = syncFilter ??
             Filter(
               room: RoomFilter(
@@ -212,10 +229,60 @@ class Client extends MatrixApi {
       EventTypes.Message,
       EventTypes.Encrypted,
       EventTypes.Sticker,
+      EventTypes.CallInvite,
+      EventTypes.CallAnswer,
+      EventTypes.CallReject,
+      EventTypes.CallHangup,
+
+      /// hack because having them both in important events and roomPreivew
+      /// makes the statekey '' which means you can only have one event of that
+      /// type
+      // EventTypes.GroupCallPrefix,
+      // EventTypes.GroupCallMemberPrefix,
     ]);
 
     // register all the default commands
     registerDefaultCommands();
+  }
+
+  /// Fetches the refreshToken from the database and tries to get a new
+  /// access token from the server and then stores it correctly. Unlike the
+  /// pure API call of `Client.refresh()` this handles the complete soft
+  /// logout case.
+  /// Throws an Exception if there is no refresh token available or the
+  /// client is not logged in.
+  Future<void> refreshAccessToken() async {
+    final storedClient = await database?.getClient(clientName);
+    final refreshToken = storedClient?.tryGet<String>('refresh_token');
+    if (refreshToken == null) {
+      throw Exception('No refresh token available');
+    }
+    final homeserverUrl = homeserver?.toString();
+    final userId = userID;
+    final deviceId = deviceID;
+    if (homeserverUrl == null || userId == null || deviceId == null) {
+      throw Exception('Cannot refresh access token when not logged in');
+    }
+
+    final tokenResponse = await refresh(refreshToken);
+
+    accessToken = tokenResponse.accessToken;
+    final expiresInMs = tokenResponse.expiresInMs;
+    final tokenExpiresAt = expiresInMs == null
+        ? null
+        : DateTime.now().add(Duration(milliseconds: expiresInMs));
+    accessTokenExpiresAt = tokenExpiresAt;
+    await database?.updateClient(
+      homeserverUrl,
+      tokenResponse.accessToken,
+      tokenExpiresAt,
+      tokenResponse.refreshToken,
+      userId,
+      deviceId,
+      deviceName,
+      prevBatch,
+      encryption?.pickledOlmAccount,
+    );
   }
 
   /// The required name for this client.
@@ -469,6 +536,7 @@ class Client extends MatrixApi {
       deviceId: deviceId,
       initialDeviceDisplayName: initialDeviceDisplayName,
       inhibitLogin: inhibitLogin,
+      refreshToken: refreshToken ?? onSoftLogout != null,
     );
 
     // Connect if there is an access token in the response.
@@ -480,8 +548,15 @@ class Client extends MatrixApi {
       throw Exception(
           'Registered but token, device ID, user ID or homeserver is null.');
     }
+    final expiresInMs = response.expiresInMs;
+    final tokenExpiresAt = expiresInMs == null
+        ? null
+        : DateTime.now().add(Duration(milliseconds: expiresInMs));
+
     await init(
         newToken: accessToken,
+        newTokenExpiresAt: tokenExpiresAt,
+        newRefreshToken: response.refreshToken,
         newUserID: userId,
         newHomeserver: homeserver,
         newDeviceName: initialDeviceDisplayName ?? '',
@@ -532,6 +607,7 @@ class Client extends MatrixApi {
       medium: medium,
       // ignore: deprecated_member_use
       address: address,
+      refreshToken: refreshToken ?? onSoftLogout != null,
     );
 
     // Connect if there is an access token in the response.
@@ -542,8 +618,16 @@ class Client extends MatrixApi {
     if (homeserver_ == null) {
       throw Exception('Registered but homerserver is null.');
     }
+
+    final expiresInMs = response.expiresInMs;
+    final tokenExpiresAt = expiresInMs == null
+        ? null
+        : DateTime.now().add(Duration(milliseconds: expiresInMs));
+
     await init(
       newToken: accessToken,
+      newTokenExpiresAt: tokenExpiresAt,
+      newRefreshToken: response.refreshToken,
       newUserID: userId,
       newHomeserver: homeserver_,
       newDeviceName: initialDeviceDisplayName ?? '',
@@ -689,6 +773,7 @@ class Client extends MatrixApi {
     CreateRoomPreset preset = CreateRoomPreset.privateChat,
     List<StateEvent>? initialState,
     Visibility? visibility,
+    HistoryVisibility? historyVisibility,
     bool waitForSync = true,
     bool groupCall = false,
     Map<String, dynamic>? powerLevelContentOverride,
@@ -703,6 +788,17 @@ class Client extends MatrixApi {
             'algorithm': supportedGroupEncryptionAlgorithms.first,
           },
           type: EventTypes.Encryption,
+        ));
+      }
+    }
+    if (historyVisibility != null) {
+      initialState ??= [];
+      if (!initialState.any((s) => s.type == EventTypes.HistoryVisibility)) {
+        initialState.add(StateEvent(
+          content: {
+            'history_visibility': historyVisibility.text,
+          },
+          type: EventTypes.HistoryVisibility,
         ));
       }
     }
@@ -980,19 +1076,19 @@ class Client extends MatrixApi {
                 []));
 
     archivedRoom.prev_batch = update.timeline?.prevBatch;
-    update.state?.forEach((event) {
-      archivedRoom.setState(Event.fromMatrixEvent(
-        event,
-        archivedRoom,
-      ));
-    });
 
-    update.timeline?.events?.forEach((event) {
-      archivedRoom.setState(Event.fromMatrixEvent(
-        event,
-        archivedRoom,
-      ));
-    });
+    final stateEvents = roomUpdate.state;
+    if (stateEvents != null) {
+      await _handleRoomEvents(archivedRoom, stateEvents, EventUpdateType.state,
+          store: false);
+    }
+
+    final timelineEvents = roomUpdate.timeline?.events;
+    if (timelineEvents != null) {
+      await _handleRoomEvents(archivedRoom, timelineEvents.reversed.toList(),
+          EventUpdateType.timeline,
+          store: false);
+    }
 
     for (var i = 0; i < timeline.events.length; i++) {
       // Try to decrypt encrypted events but don't update the database.
@@ -1286,6 +1382,8 @@ class Client extends MatrixApi {
       accessToken = token;
     }
 
+    await ensureNotSoftLoggedOut();
+
     // Check if the notification contains an event at all:
     final eventId = notification.eventId;
     final roomId = notification.roomId;
@@ -1446,6 +1544,8 @@ class Client extends MatrixApi {
   /// `userDeviceKeysLoading` where it is necessary.
   Future<void> init({
     String? newToken,
+    DateTime? newTokenExpiresAt,
+    String? newRefreshToken,
     Uri? newHomeserver,
     String? newUserID,
     String? newDeviceName,
@@ -1481,7 +1581,7 @@ class Client extends MatrixApi {
     String? userID;
     try {
       Logs().i('Initialize client $clientName');
-      if (isLogged()) {
+      if (onLoginStateChanged.value == LoginState.loggedIn) {
         throw ClientInitPreconditionError(
           'User is already logged in! Call [logout()] first!',
         );
@@ -1499,10 +1599,16 @@ class Client extends MatrixApi {
       _serverConfigCache.invalidate();
 
       final account = await this.database?.getClient(clientName);
+      newRefreshToken ??= account?.tryGet<String>('refresh_token');
       if (account != null) {
         _id = account['client_id'];
         homeserver = Uri.parse(account['homeserver_url']);
         accessToken = this.accessToken = account['token'];
+        final tokenExpiresAtMs =
+            int.tryParse(account.tryGet<String>('token_expires_at') ?? '');
+        accessTokenExpiresAt = tokenExpiresAtMs == null
+            ? null
+            : DateTime.fromMillisecondsSinceEpoch(tokenExpiresAtMs);
         userID = _userID = account['user_id'];
         _deviceID = account['device_id'];
         _deviceName = account['device_name'];
@@ -1512,6 +1618,7 @@ class Client extends MatrixApi {
       }
       if (newToken != null) {
         accessToken = this.accessToken = newToken;
+        accessTokenExpiresAt = newTokenExpiresAt;
         homeserver = newHomeserver;
         userID = _userID = newUserID;
         _deviceID = newDeviceID;
@@ -1519,11 +1626,32 @@ class Client extends MatrixApi {
         olmAccount = newOlmAccount;
       } else {
         accessToken = this.accessToken = newToken ?? accessToken;
+        accessTokenExpiresAt = newTokenExpiresAt ?? accessTokenExpiresAt;
         homeserver = newHomeserver ?? homeserver;
         userID = _userID = newUserID ?? userID;
         _deviceID = newDeviceID ?? _deviceID;
         _deviceName = newDeviceName ?? _deviceName;
         olmAccount = newOlmAccount ?? olmAccount;
+      }
+
+      // If we are refreshing the session, we are done here:
+      if (onLoginStateChanged.value == LoginState.softLoggedOut) {
+        if (newRefreshToken != null && accessToken != null && userID != null) {
+          // Store the new tokens:
+          await _database?.updateClient(
+            homeserver.toString(),
+            accessToken,
+            accessTokenExpiresAt,
+            newRefreshToken,
+            userID,
+            _deviceID,
+            _deviceName,
+            prevBatch,
+            encryption?.pickledOlmAccount,
+          );
+        }
+        onLoginStateChanged.add(LoginState.loggedIn);
+        return;
       }
 
       if (accessToken == null || homeserver == null || userID == null) {
@@ -1559,6 +1687,8 @@ class Client extends MatrixApi {
           await database.updateClient(
             homeserver.toString(),
             accessToken,
+            accessTokenExpiresAt,
+            newRefreshToken,
             userID,
             _deviceID,
             _deviceName,
@@ -1570,6 +1700,8 @@ class Client extends MatrixApi {
             clientName,
             homeserver.toString(),
             accessToken,
+            accessTokenExpiresAt,
+            newRefreshToken,
             userID,
             _deviceID,
             _deviceName,
@@ -1663,6 +1795,10 @@ class Client extends MatrixApi {
   bool get syncPending => _currentSync != null;
 
   /// Controls the background sync (automatically looping forever if turned on).
+  /// If you use soft logout, you need to manually call
+  /// `ensureNotSoftLoggedOut()` before doing any API request after setting
+  /// the background sync to false, as the soft logout is handeld automatically
+  /// in the sync loop.
   set backgroundSync(bool enabled) {
     _backgroundSync = enabled;
     if (_backgroundSync) {
@@ -1702,6 +1838,34 @@ class Client extends MatrixApi {
     return;
   }
 
+  Future<void> _handleSoftLogout() async {
+    final onSoftLogout = this.onSoftLogout;
+    if (onSoftLogout == null) return;
+
+    onLoginStateChanged.add(LoginState.softLoggedOut);
+    try {
+      await onSoftLogout(this);
+      onLoginStateChanged.add(LoginState.loggedIn);
+    } catch (e, s) {
+      Logs().w('Unable to refresh session after soft logout', e, s);
+      await clear();
+      rethrow;
+    }
+  }
+
+  /// Checks if the token expires in under [expiresIn] time and calls the
+  /// given `onSoftLogout()` if so. You have to provide `onSoftLogout` in the
+  /// Client constructor. Otherwise this will do nothing.
+  Future<void> ensureNotSoftLoggedOut(
+      [Duration expiresIn = const Duration(minutes: 1)]) async {
+    final tokenExpiresAt = accessTokenExpiresAt;
+    if (onSoftLogout != null &&
+        tokenExpiresAt != null &&
+        tokenExpiresAt.difference(DateTime.now()) <= expiresIn) {
+      await _handleSoftLogout();
+    }
+  }
+
   /// Pass a timeout to set how long the server waits before sending an empty response.
   /// (Corresponds to the timeout param on the /sync request.)
   Future<void> _innerSync({Duration? timeout}) async {
@@ -1720,6 +1884,8 @@ class Client extends MatrixApi {
       // server that we want to receive an empty sync response after this
       // amount of time if nothing happens.
       timeout ??= const Duration(seconds: 30);
+
+      await ensureNotSoftLoggedOut(timeout * 2);
 
       final syncRequest = sync(
         filter: syncFilterId,
@@ -1788,16 +1954,21 @@ class Client extends MatrixApi {
         await processToDeviceQueue();
       } catch (_) {} // we want to dispose any errors this throws
 
-      await singleShotStaleCallChecker();
-
       _retryDelay = Future.value();
       onSyncStatus.add(SyncStatusUpdate(SyncStatus.finished));
     } on MatrixException catch (e, s) {
       onSyncStatus.add(SyncStatusUpdate(SyncStatus.error,
           error: SdkError(exception: e, stackTrace: s)));
       if (e.error == MatrixError.M_UNKNOWN_TOKEN) {
-        Logs().w('The user has been logged out!');
-        await clear();
+        final onSoftLogout = this.onSoftLogout;
+        if (e.raw.tryGet<bool>('soft_logout') == true && onSoftLogout != null) {
+          Logs().w('The user has been soft logged out! Try to login again...');
+
+          await _handleSoftLogout();
+        } else {
+          Logs().w('The user has been logged out!');
+          await clear();
+        }
       }
     } on SyncConnectionException catch (e, s) {
       Logs().w('Syncloop failed: Client has not connection to the server');
@@ -1815,7 +1986,9 @@ class Client extends MatrixApi {
   /// Use this method only for testing utilities!
   Future<void> handleSync(SyncUpdate sync, {Direction? direction}) async {
     // ensure we don't upload keys because someone forgot to set a key count
-    sync.deviceOneTimeKeysCount ??= {'signed_curve25519': 100};
+    sync.deviceOneTimeKeysCount ??= {
+      'signed_curve25519': encryption?.olmManager.maxNumberOfOneTimeKeys ?? 100,
+    };
     await _handleSync(sync, direction: direction);
   }
 
@@ -2339,6 +2512,10 @@ class Client extends MatrixApi {
           final importantOrRoomLoaded =
               eventUpdate.type == EventUpdateType.inviteState ||
                   !room.partial ||
+                  // make sure we do overwrite events we have already loaded.
+                  room.states[stateEvent.type]
+                          ?.containsKey(stateEvent.stateKey ?? '') ==
+                      true ||
                   importantStateEvents.contains(stateEvent.type);
           if ((noMessageOrNoEdit || editingLastEvent || consecutiveEdit) &&
               importantOrRoomLoaded) {
@@ -3080,10 +3257,16 @@ class Client extends MatrixApi {
     Logs().i('Found data in the legacy database!');
     onMigration?.call();
     _id = migrateClient['client_id'];
+    final tokenExpiresAtMs =
+        int.tryParse(migrateClient.tryGet<String>('token_expires_at') ?? '');
     await database.insertClient(
       clientName,
       migrateClient['homeserver_url'],
       migrateClient['token'],
+      tokenExpiresAtMs == null
+          ? null
+          : DateTime.fromMillisecondsSinceEpoch(tokenExpiresAtMs),
+      migrateClient['refresh_token'],
       migrateClient['user_id'],
       migrateClient['device_id'],
       migrateClient['device_name'],