+++ lib/src/database/hive_database.dart
@@ -28,6 +28,7 @@ import 'package:matrix/encryption/utils/outbound_group_session.dart';
 import 'package:matrix/encryption/utils/ssss_cache.dart';
 import 'package:matrix/encryption/utils/stored_inbound_group_session.dart';
 import 'package:matrix/matrix.dart';
+import 'package:matrix/src/database/zone_transaction_mixin.dart';
 import 'package:matrix/src/utils/copy_map.dart';
 import 'package:matrix/src/utils/queued_to_device_event.dart';
 import 'package:matrix/src/utils/run_benchmarked.dart';
@@ -39,7 +40,7 @@ import 'package:matrix/src/utils/run_benchmarked.dart';
 /// This database does not support file caching!
 @Deprecated(
     'Use [HiveCollectionsDatabase] instead. Don\'t forget to properly migrate!')
-class FamedlySdkHiveDatabase extends DatabaseApi {
+class FamedlySdkHiveDatabase extends DatabaseApi with ZoneTransactionMixin {
   static const int version = 5;
   final String name;
   late Box _clientBox;
@@ -749,6 +750,8 @@ class FamedlySdkHiveDatabase extends DatabaseApi {
       String name,
       String homeserverUrl,
       String token,
+      DateTime? tokenExpiresAt,
+      String? refreshToken,
       String userId,
       String? deviceId,
       String? deviceName,
@@ -756,6 +759,9 @@ class FamedlySdkHiveDatabase extends DatabaseApi {
       String? olmAccount) async {
     await _clientBox.put('homeserver_url', homeserverUrl);
     await _clientBox.put('token', token);
+    await _clientBox.put(
+        'token_expires_at', tokenExpiresAt?.millisecondsSinceEpoch.toString());
+    await _clientBox.put('refresh_token', refreshToken);
     await _clientBox.put('user_id', userId);
     await _clientBox.put('device_id', deviceId);
     await _clientBox.put('device_name', deviceName);
@@ -1305,62 +1311,16 @@ class FamedlySdkHiveDatabase extends DatabaseApi {
     return;
   }
 
-  Completer<void>? _transactionLock;
-  final _transactionZones = <Zone>{};
-
   @override
-  Future<void> transaction(Future<void> Function() action) async {
-    // we want transactions to lock, however NOT if transactoins are run inside of each other.
-    // to be able to do this, we use dart zones (https://dart.dev/articles/archive/zones).
-    // _transactionZones holds a set of all zones which are currently running a transaction.
-    // _transactionLock holds the lock.
-
-    // first we try to determine if we are inside of a transaction currently
-    var isInTransaction = false;
-    Zone? zone = Zone.current;
-    // for that we keep on iterating to the parent zone until there is either no zone anymore
-    // or we have found a zone inside of _transactionZones.
-    while (zone != null) {
-      if (_transactionZones.contains(zone)) {
-        isInTransaction = true;
-        break;
-      }
-      zone = zone.parent;
-    }
-    // if we are inside a transaction....just run the action
-    if (isInTransaction) {
-      return await action();
-    }
-    // if we are *not* in a transaction, time to wait for the lock!
-    while (_transactionLock != null) {
-      await _transactionLock!.future;
-    }
-    // claim the lock
-    final lock = Completer<void>();
-    _transactionLock = lock;
-    try {
-      // run the action inside of a new zone
-      return await runZoned(() async {
-        try {
-          // don't forget to add the new zone to _transactionZones!
-          _transactionZones.add(Zone.current);
-          return await action();
-        } finally {
-          // aaaand remove the zone from _transactionZones again
-          _transactionZones.remove(Zone.current);
-        }
-      });
-    } finally {
-      // aaaand finally release the lock
-      _transactionLock = null;
-      lock.complete();
-    }
-  }
+  Future<void> transaction(Future<void> Function() action) =>
+      zoneTransaction(action);
 
   @override
   Future<void> updateClient(
     String homeserverUrl,
     String token,
+    DateTime? tokenExpiresAt,
+    String? refreshToken,
     String userId,
     String? deviceId,
     String? deviceName,
@@ -1369,6 +1329,9 @@ class FamedlySdkHiveDatabase extends DatabaseApi {
   ) async {
     await _clientBox.put('homeserver_url', homeserverUrl);
     await _clientBox.put('token', token);
+    await _clientBox.put(
+        'token_expires_at', tokenExpiresAt?.millisecondsSinceEpoch.toString());
+    await _clientBox.put('refresh_token', refreshToken);
     await _clientBox.put('user_id', userId);
     await _clientBox.put('device_id', deviceId);
     await _clientBox.put('device_name', deviceName);