+++ lib/src/database/matrix_sdk_database.dart
@@ -428,10 +428,11 @@ class MatrixSdkDatabase extends DatabaseApi {
         // Combine those two lists while respecting the start and limit parameters.
         final end = min(timelineEventIds.length,
             start + (limit ?? timelineEventIds.length));
-        final eventIds = sendingEventIds +
-            (start < timelineEventIds.length && !onlySending
-                ? timelineEventIds.getRange(start, end).toList()
-                : []);
+        final eventIds = [
+          ...sendingEventIds,
+          if (!onlySending && start < timelineEventIds.length)
+            ...timelineEventIds.getRange(start, end),
+        ];
 
         return await _getEventsByIds(eventIds.cast<String>(), room);
       });
@@ -726,6 +727,8 @@ class MatrixSdkDatabase extends DatabaseApi {
       String name,
       String homeserverUrl,
       String token,
+      DateTime? tokenExpiresAt,
+      String? refreshToken,
       String userId,
       String? deviceId,
       String? deviceName,
@@ -734,6 +737,17 @@ class MatrixSdkDatabase extends DatabaseApi {
     await transaction(() async {
       await _clientBox.put('homeserver_url', homeserverUrl);
       await _clientBox.put('token', token);
+      if (tokenExpiresAt == null) {
+        await _clientBox.delete('token_expires_at');
+      } else {
+        await _clientBox.put('token_expires_at',
+            tokenExpiresAt.millisecondsSinceEpoch.toString());
+      }
+      if (refreshToken == null) {
+        await _clientBox.delete('refresh_token');
+      } else {
+        await _clientBox.put('refresh_token', refreshToken);
+      }
       await _clientBox.put('user_id', userId);
       if (deviceId == null) {
         await _clientBox.delete('device_id');
@@ -1342,6 +1356,8 @@ class MatrixSdkDatabase extends DatabaseApi {
   Future<void> updateClient(
     String homeserverUrl,
     String token,
+    DateTime? tokenExpiresAt,
+    String? refreshToken,
     String userId,
     String? deviceId,
     String? deviceName,
@@ -1351,6 +1367,17 @@ class MatrixSdkDatabase extends DatabaseApi {
     await transaction(() async {
       await _clientBox.put('homeserver_url', homeserverUrl);
       await _clientBox.put('token', token);
+      if (tokenExpiresAt == null) {
+        await _clientBox.delete('token_expires_at');
+      } else {
+        await _clientBox.put('token_expires_at',
+            tokenExpiresAt.millisecondsSinceEpoch.toString());
+      }
+      if (refreshToken == null) {
+        await _clientBox.delete('refresh_token');
+      } else {
+        await _clientBox.put('refresh_token', refreshToken);
+      }
       await _clientBox.put('user_id', userId);
       if (deviceId == null) {
         await _clientBox.delete('device_id');