+++ test/event_test.dart
@@ -1089,6 +1089,57 @@ void main() {
               plaintextBody: true,
               withSenderNamePrefix: true),
           'Example: Title\nsome text and 🔗link\nokay and this is important');
+
+      event = Event.fromJson({
+        'content': {
+          'body':
+              'Alice is requesting to verify your device, but your client does not support verification, so you may need to use a different verification method.',
+          'from_device': 'AliceDevice2',
+          'methods': ['m.sas.v1'],
+          'msgtype': 'm.key.verification.request',
+          'to': '@bob:example.org'
+        },
+        'event_id': '\$143273582443PhrSn:example.org',
+        'origin_server_ts': 1432735824653,
+        'room_id': '!jEsUZKDJdhlrceRyVU:example.org',
+        'sender': '@example:example.org',
+        'type': 'm.room.message',
+        'unsigned': {'age': 1234}
+      }, room);
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example requested key verification',
+      );
+
+      event.content['msgtype'] = 'm.key.verification.ready';
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example is ready for key verification',
+      );
+
+      event.content['msgtype'] = 'm.key.verification.start';
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example started key verification',
+      );
+
+      event.content['msgtype'] = 'm.key.verification.cancel';
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example canceled key verification',
+      );
+
+      event.content['msgtype'] = 'm.key.verification.done';
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example completed key verification',
+      );
+
+      event.content['msgtype'] = 'm.key.verification.accept';
+      expect(
+        await event.calcLocalizedBody(MatrixDefaultLocalizations()),
+        'Example accepted key verification request',
+      );
     });
 
     test('aggregations', () {
@@ -1498,6 +1549,93 @@ void main() {
 
       await room.client.dispose(closeDatabase: true);
     });
+
+    test('downloadAndDecryptAttachment store only', () async {
+      final FILE_BUFF = Uint8List.fromList([0]);
+      var serverHits = 0;
+      Future<Uint8List> downloadCallback(Uri uri) async {
+        serverHits++;
+        return {
+          '/_matrix/media/v3/download/example.org/newfile': FILE_BUFF,
+        }[uri.path]!;
+      }
+
+      await client.checkHomeserver(Uri.parse('https://fakeserver.notexisting'),
+          checkWellKnown: false);
+      final room = Room(id: '!localpart:server.abc', client: await getClient());
+      final event = Event.fromJson({
+        'type': EventTypes.Message,
+        'content': {
+          'body': 'image',
+          'msgtype': 'm.image',
+          'url': 'mxc://example.org/newfile',
+          'info': {
+            'size': 5,
+          },
+        },
+        'event_id': '\$edit2',
+        'sender': '@alice:example.org',
+      }, room);
+
+      var buffer = await event.downloadAndDecryptAttachment(
+          downloadCallback: downloadCallback);
+      expect(await event.isAttachmentInLocalStore(),
+          event.room.client.database?.supportsFileStoring);
+      expect(buffer.bytes, FILE_BUFF);
+      expect(serverHits, 1);
+
+      if (event.room.client.database?.supportsFileStoring == true) {
+        buffer = await event.downloadAndDecryptAttachment(
+            downloadCallback: downloadCallback, fromLocalStoreOnly: true);
+        expect(buffer.bytes, FILE_BUFF);
+      } else {
+        expect(
+            () async => await event.downloadAndDecryptAttachment(
+                downloadCallback: downloadCallback, fromLocalStoreOnly: true),
+            throwsA(anything));
+      }
+      expect(serverHits, 1);
+
+      await room.client.dispose(closeDatabase: true);
+    });
+
+    test('downloadAndDecryptAttachment store only without file', () async {
+      final FILE_BUFF = Uint8List.fromList([0]);
+      var serverHits = 0;
+      Future<Uint8List> downloadCallback(Uri uri) async {
+        serverHits++;
+        return {
+          '/_matrix/media/v3/download/example.org/newfile': FILE_BUFF,
+        }[uri.path]!;
+      }
+
+      await client.checkHomeserver(Uri.parse('https://fakeserver.notexisting'),
+          checkWellKnown: false);
+      final room = Room(id: '!localpart:server.abc', client: await getClient());
+      final event = Event.fromJson({
+        'type': EventTypes.Message,
+        'content': {
+          'body': 'image',
+          'msgtype': 'm.image',
+          'url': 'mxc://example.org/newfile',
+          'info': {
+            'size': 5,
+          },
+        },
+        'event_id': '\$edit2',
+        'sender': '@alice:example.org',
+      }, room);
+
+      expect(
+          () async => await event.downloadAndDecryptAttachment(
+              downloadCallback: downloadCallback, fromLocalStoreOnly: true),
+          throwsA(anything));
+
+      expect(serverHits, 0);
+
+      await room.client.dispose(closeDatabase: true);
+    });
+
     test('emote detection', () async {
       var event = Event.fromJson({
         'type': EventTypes.Message,